major restructuring of partial product compressor classes, documentation fixes, new ones complement adder
