+++ doc/components/adder.md
@@ -1,16 +1,17 @@
 # Adder
 
-ROHD HCL provides a set of adder modules to get the sum from a list of logic. As of now, we have
+ROHD-HCL provides a set of adder modules to get the sum from a pair of Logic. As of now, we have
 
 - [Ripple Carry Adder](#ripple-carry-adder)
 - [Parallel Prefix Adder](#parallel-prefix-adder)
+- [One's Complement Adder Subtractor](#ones-complement-adder-subtractor)
 - [Sign Magnitude Adder](#sign-magnitude-adder)
 
 ## Ripple Carry Adder
 
 A ripple carry adder is a digital circuit used for binary addition. It consists of a series of full adders connected in a chain, with the carry output of each adder linked to the carry input of the next one. Starting from the least significant bit (LSB) to most significant bit (MSB), the adder sequentially adds corresponding bits of two binary numbers.
 
-The [`RippleCarryAdder`](https://intel.github.io/rohd-hcl/rohd_hcl/RippleCarryAdder-class.html) module in ROHD-HCL accept input  `Logic`s a and b as the input pin and the name of the module `name`. Note that the width of the inputs must be the same or a `RohdHclException` will be thrown.
+The [`RippleCarryAdder`](https://intel.github.io/rohd-hcl/rohd_hcl/RippleCarryAdder-class.html) module in ROHD-HCL accept input `Logic`s a and b as the input pin and the name of the module `name`. Note that the width of the inputs must be the same or a `RohdHclException` will be thrown.
 
 An example is shown below to add two inputs of signals that have 8-bits of width.
 
@@ -27,7 +28,7 @@ final sum = rippleCarryAdder.sum;
 
 ## Parallel Prefix Adder
 
-A parallel prefix adder is an adder that uses n instance of a parallel prefix tree (see `Parallel Prefix Operations`) to efficiently connect a set of `Full Adder` circuits to form a complete adder.
+A parallel prefix adder is an adder that uses different varieties of a parallel prefix tree (see `Parallel Prefix Operations`) to efficiently connect a set of `Full Adder` circuits to form a complete adder.
 
 Here is an example of instantiating a `ParallelPrefixAdder`:
 
@@ -46,11 +47,39 @@ Here is an example of instantiating a `ParallelPrefixAdder`:
     print('${sum.value.toBigInt()}');
 ```
 
+## One's Complement Adder Subtractor
+
+A ones-complement adder (and subtractor) is useful in efficient arithmetic operations as the
+end-around carry can be bypassed and used later.
+
+The `OnesComplementAdder` can take a subtraction command as either a Logic `subtractIn` or a boolean `subtract` (the Logic overrides the boolean).  If Logic `carry` is provided, the end-around carry is output on `carry` and the value will be one less than expected when `carry` is high.  An `adderGen` adder function can be provided that generates your favorite internal adder (such as a parallel prefix adder).
+
+The output of `OnesComplementAdder` is a `sum` which is the magnitude and a `sign`.
+
+Here is an example of instantiating a `OnesComplementAdder` as a subtractor, but saving the `carry`:
+
+```dart
+    const width = 4;
+    final a = Logic(width: width);
+    final b = Logic(width: width);
+
+    a.put(av);
+    b.put(bv);
+    final carry = Logic();
+    final adder = OnesComplementAdder(
+        a, b, null, carry, RippleCarryAdder.new,
+        subtract: true);
+    final mag = adder.sum.value.toInt() + (carry.value.isZero ? 0 : 1));
+    final out = (adder.sign.value.toInt() == 1 ? -mag : mag);
+```
+
 ## Sign Magnitude Adder
 
-A sign magnitude adder is useful in situations where the sign of the addends is seperated from their magnitude (e.g., not 2s complement), such as in floating point multipliers.  The `SignMagnitudeAdder` inherits from `Adder` but adds the `Logic` inputs for the two operands.
+A sign magnitude adder is useful in situations where the sign of the addends is separated from their magnitude (e.g., not 2s complement), such as in floating point multipliers.  The `SignMagnitudeAdder` inherits from `Adder` but adds the `Logic` inputs for the two operands.
 
-If you can supply the largest magnitude number first, then you can disable a comparator generation inside by declaring the  `largestMagnitudeFirst` option as true.
+If you can supply the largest magnitude number first, then you can disable a comparator generation inside by declaring the `largestMagnitudeFirst` option as true.
+
+The `SignMagnitudeAdder` uses a `OnesComplementAdder` internally.
 
 Here is an example of instantiating a `SignMagnitudeAdder`:
