+++ lib/rohd_hcl.dart
@@ -3,12 +3,7 @@
 
 export 'src/arbiters/arbiters.dart';
 export 'src/arithmetic/adder.dart';
-export 'src/arithmetic/carry_save_mutiplier.dart';
-export 'src/arithmetic/multiplier.dart';
-export 'src/arithmetic/multiplier_lib.dart';
-export 'src/arithmetic/parallel_prefix_operations.dart';
-export 'src/arithmetic/ripple_carry_adder.dart';
-export 'src/arithmetic/sign_magnitude_adder.dart';
+export 'src/arithmetic/arithmetic.dart';
 export 'src/binary_gray.dart';
 export 'src/component_config/component_config.dart';
 export 'src/count.dart';
