+++ lib/src/arithmetic/sign_magnitude_adder.dart
@@ -11,52 +11,109 @@ import 'package:meta/meta.dart';
 import 'package:rohd/rohd.dart';
 import 'package:rohd_hcl/rohd_hcl.dart';
 
-/// An Adder which performs one's complement arithmetic using an
-/// adder that is passed in using a functor. If the caller can guarantee
-/// that the larger magnitude or negative value is provided first in 'a', then
-/// they can set 'largestMagnitudeFirst' to 'true' to avoid a comparator.
+/// A [SignMagnitudeAdder] performsa addition on values in sign/magnitude format.
 class SignMagnitudeAdder extends Adder {
   /// The sign of the first input
-  @protected
-  late final Logic aSign;
+  Logic aSign;
 
   /// The sign of the second input
-  @protected
-  late final Logic bSign;
+  Logic bSign;
 
   /// The sign of the result
   Logic get sign => output('sign');
+
+  @protected
+  Logic _sign = Logic();
 
   /// Largest magnitude argument is provided in [a] or if equal
   /// the argument with a negative sign.
   bool largestMagnitudeFirst;
 
-  /// [SignMagnitudeAdder] constructor with an unsigned adder functor
-  SignMagnitudeAdder(Logic as, super.a, Logic bs, super.b,
+  /// [SignMagnitudeAdder] constructor with an adder functor [adderGen]
+  ///Inputs are (sign, magnitude) pairs: ([aSign], [a]) and ([bSign], [b]).
+  /// If the caller can guarantee that the larger magnitude value
+  ///  is provided first in [a], then they can set [largestMagnitudeFirst]
+  /// too 'true' to avoid a comparator.
+  // TODO(desmonddak): this adder may need a carry-in for rounding
+  SignMagnitudeAdder(this.aSign, super.a, this.bSign, super.b,
       Adder Function(Logic, Logic) adderGen,
       {this.largestMagnitudeFirst = false})
       : super(
-            name: 'Ones Complement Adder: '
+            name: 'Sign Magnitude Adder: '
                 '${adderGen.call(Logic(), Logic()).name}') {
-    aSign = addInput('aSign', as);
-    bSign = addInput('bSign', bs);
-    final sign = addOutput('sign');
+    aSign = addInput('aSign', aSign);
+    bSign = addInput('bSign', bSign);
+    _sign = addOutput('sign');
 
     final bLarger = a.lt(b) | (a.eq(b) & bSign.gt(aSign));
 
-    final computeSign = mux(largestMagnitudeFirst ? Const(1) : Const(0), aSign,
-        mux(bLarger, bSign, aSign));
+    _sign <= (largestMagnitudeFirst ? aSign : mux(bLarger, bSign, aSign));
+    final adder = OnesComplementAdder(a, b, aSign ^ bSign, null, adderGen);
+    sum <= adder.sum;
+  }
+}
 
-    final ax = a.zeroExtend(a.width + 1);
-    final bx = b.zeroExtend(b.width + 1);
+/// An adder (and subtractor) [OnesComplementAdder] that operates on
+/// ones-complement values.
+class OnesComplementAdder extends Adder {
+  /// The sign of the result
+  Logic get sign => output('sign');
 
-    final aOnesComplement = mux(aSign, ~ax, ax);
-    final bOnesComplement = mux(bSign, ~bx, bx);
+  /// The end-around carry which should be added to the resulting [sum]
+  /// If the input [carry] is not null, this value is stored there. Otherwise,
+  /// the end-around carry is internally added to [sum]
+  Logic? get carry => tryOutput('carry');
 
-    final adder = adderGen(aOnesComplement, bOnesComplement);
-    final endAround = adder.sum[-1] & (aSign | bSign);
-    final localOut = mux(endAround, adder.sum + 1, adder.sum);
-    sign <= computeSign;
-    sum <= mux(sign, ~localOut, localOut).slice(ax.width - 1, 0);
+  @protected
+  Logic _sign = Logic();
+
+  /// [OnesComplementAdder] constructor with an adder functor [adderGen]
+  /// Either a Logic [subtractIn] or a boolean [subtract] can enable
+  /// subtraction, with [subtractIn] overriding [subtract].  If Logic [carry]
+  /// is provided as not null, then the end-around carry is not performed and is
+  /// left to the caller via the output [carry].
+  OnesComplementAdder(super.a, super.b, Logic? subtractIn, Logic? carry,
+      Adder Function(Logic, Logic) adderGen,
+      {bool subtract = false})
+      : super(
+            name: 'Ones Complement Adder: '
+                '${adderGen.call(Logic(), Logic()).name}') {
+    if (subtractIn != null) {
+      subtractIn = addInput('subtractIn', subtractIn);
+    }
+    _sign = addOutput('sign');
+    if (carry != null) {
+      addOutput('carry');
+      carry <= this.carry!;
+    }
+    if ((subtractIn != null) & subtract) {
+      throw RohdHclException(
+          'Subtraction is controlled by a non-null subtractIn: '
+          'subtract boolean is ignored');
+    }
+    final doSubtract = subtractIn ?? (subtract ? Const(1) : Const(0));
+
+    final ax = a.zeroExtend(a.width);
+    final bx = b.zeroExtend(b.width);
+
+    final adder = adderGen(ax, mux(doSubtract, ~bx, bx));
+
+    if (this.carry != null) {
+      this.carry! <= adder.sum[-1];
+    }
+    final endAround = mux(doSubtract, adder.sum[-1], Const(0));
+    final magnitude = adder.sum.slice(a.width - 1, 0);
+
+    sum <=
+        mux(
+            doSubtract,
+            mux(
+                    endAround,
+                    [if (this.carry != null) magnitude else magnitude + 1]
+                        .first,
+                    ~magnitude)
+                .zeroExtend(sum.width),
+            adder.sum);
+    _sign <= mux(doSubtract, ~endAround, Const(0));
   }
 }
